refactor: update file and folder names to kebab-case
Converted file and folder naming conventions from PascalCase to kebab-case for consistency. Adjusted all relevant imports and exports to reflect these changes.
