--- apps/nextjs/src/components/pages/home/components/Content/hooks/index.ts
@@ -1,2 +1,0 @@
-export { useHomeContentDom } from './useHomeContentDom';
-export { useTasksListContentVerticalScroll } from './useHomeContentVerticalScroll';
