--- apps/nextjs/src/components/pages/home/components/FavoriteProjects/index.ts
@@ -1,1 +1,0 @@
-export { FavoriteProjects } from './FavoriteProjects';
