--- apps/nextjs/src/components/pages/home/components/ProjectsContainer/index.ts
@@ -1,1 +1,0 @@
-export { PADDING_X, ProjectsContainer } from './ProjectsContainer';
