--- apps/nextjs/src/components/pages/home/components/RecentProjects/index.ts
@@ -1,1 +1,0 @@
-export { RecentProjects } from './RecentProjects';
