--- apps/nextjs/src/components/pages/home/components/TasksDueSoon/index.ts
@@ -1,1 +1,0 @@
-export { TasksDueSoon } from './TasksDueSoon';
