+++ apps/nextjs/src/components/pages/home/components/content/content.tsx
