+++ apps/nextjs/src/components/pages/home/components/content/hooks/index.ts
@@ -1,0 +1,2 @@
+export { useHomeContentDom } from './use-home-content-dom';
+export { useTasksListContentVerticalScroll } from './use-home-content-vertical-scroll';
