+++ apps/nextjs/src/components/pages/home/components/content/hooks/use-home-content-dom.ts
