+++ apps/nextjs/src/components/pages/home/components/content/index.ts
@@ -1,2 +1,2 @@
-export { Content } from './Content';
+export { Content } from './content';
 export { useHomeContentDom, useTasksListContentVerticalScroll } from './hooks';
