+++ apps/nextjs/src/components/pages/home/components/favorite-projects/favorite-projects.tsx
@@ -1,5 +1,5 @@
 import { memo } from 'react';
-import { ProjectsContainer } from '@/components/pages/home/components/ProjectsContainer';
+import { ProjectsContainer } from '@/components/pages/home/components/projects-container';
 import { useFavoriteProjectIds } from '@/store/entities/favoriteProjectIds';
 
 export const FavoriteProjects = memo(function FavoriteProjects() {
