+++ apps/nextjs/src/components/pages/home/components/favorite-projects/index.ts
@@ -1,0 +1,1 @@
+export { FavoriteProjects } from './favorite-projects';
