+++ apps/nextjs/src/components/pages/home/components/header/index.ts
@@ -1,0 +1,1 @@
+export { Header } from './header';
