+++ apps/nextjs/src/components/pages/home/components/projects-container/index.ts
@@ -1,0 +1,1 @@
+export { PADDING_X, ProjectsContainer } from './projects-container';
