+++ apps/nextjs/src/components/pages/home/components/projects-container/projects-container.tsx
