+++ apps/nextjs/src/components/pages/home/components/recent-projects/index.ts
@@ -1,0 +1,1 @@
+export { RecentProjects } from './recent-projects';
