+++ apps/nextjs/src/components/pages/home/components/skeleton-home.tsx
