+++ apps/nextjs/src/components/pages/home/components/tasks-due-soon/index.ts
@@ -1,0 +1,1 @@
+export { TasksDueSoon } from './tasks-due-soon';
