+++ apps/nextjs/src/components/pages/home/index.ts
@@ -1,1 +1,1 @@
-export { Container } from './Container';
+export { Container } from './container';
